packages/agent-sre/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-sre"
version = "3.0.0"
description = "Public Preview — Reliability Engineering for AI Agent Systems"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Microsoft Corporation", email = "agentgovtoolkit@microsoft.com"},
]
maintainers = [
    {name = "Agent Governance Toolkit Team", email = "agentgovtoolkit@microsoft.com"},
]
keywords = ["ai", "agents", "sre", "reliability", "observability", "slo"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "pydantic>=2.4.0",
    "pyyaml>=6.0",
    "croniter>=2.0",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.8",
    "mypy>=1.10",
]
api = ["fastapi>=0.115.0", "uvicorn>=0.27.0"]
otel = [
    "opentelemetry-exporter-otlp>=1.20",
]
langfuse = ["langfuse>=2.0"]
arize = ["arize-phoenix>=4.0"]
langchain = ["langchain-core>=1.2.11"]
llamaindex = ["llama-index-core>=0.10"]
braintrust = ["braintrust>=0.0.100"]
helicone = ["helicone>=4.0"]
datadog = ["ddtrace>=2.0"]
langsmith = ["langsmith>=0.1"]
wandb = ["wandb>=0.16"]
mlflow = ["mlflow>=2.10"]
agentops = ["agentops>=0.3"]
all = ["opentelemetry-exporter-otlp>=1.20", "langfuse>=2.0", "arize-phoenix>=4.0", "langchain-core>=1.2.11", "llama-index-core>=0.10", "braintrust>=0.0.100", "helicone>=4.0", "ddtrace>=2.0", "langsmith>=0.1", "wandb>=0.16", "mlflow>=2.10", "agentops>=0.3"]

[project.scripts]
agent-sre = "agent_sre.cli.main:cli"

[project.urls]
Homepage = "https://github.com/microsoft/agent-governance-toolkit"
Repository = "https://github.com/microsoft/agent-governance-toolkit"
Issues = "https://github.com/microsoft/agent-governance-toolkit/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_sre"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "TCH"]
ignore = [
    "E501",   # line too long
    "E741",   # ambiguous variable name
    "UP035",  # deprecated typing imports (need py310 compat)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "E402"]
"tests/*" = ["F401", "E402"]
"src/agent_sre/integrations/*" = ["E402"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
packages = ["agent_sre"]
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: long-running tests (skip with -m 'not slow')",
    "integration: end-to-end integration tests",
]

[tool.bandit]
skips = ["B310"]
